Complete basic library function feature and server-client connection
