+++ packages/keyda-demo/src/index.js
@@ -1,13 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
-import { KeyStateProvider } from './lib/Context';
 
 ReactDOM.render(
   <React.StrictMode>
-    <KeyStateProvider>
-      <App />
-    </KeyStateProvider>
+    <App />
   </React.StrictMode>,
   document.getElementById('root')
 );
